feat: add Tailwind CSS configuration for React components
this commit includes creating a PostCSS configuration file and a Tailwind CSS configuration file. New styles are introduced, including the specific configuration for the Roboto font and the addition of a "darkMode" option. Various dependency updates in the package-lock.json file are also made as necessary. This addition of Tailwind CSS allows for a more streamlined styling workflow in the React libraries.