feat: this commit adds a new library named `shareable`.
It also includes all related configuration files, such as eslint, jest config, tsconfig, package.json, and others. The `shareable` library is a simple library that provides a single function which returns a 'shareable' string. The code also includes a spec file for testing this function. Alongside the setup for the library, the tsconfig paths and the package-lock.json file were updated to include the new shareable library.