feat: add new button style, update components to use 'shareable' library
this commit adds the implementation of a new button style and completes the migration of utility functions to the 'shareable' library. The functionality of all components have been updated to use the utilities from 'shareable'. The version of '@vitejs/plugin-react' in package-lock.json has also been updated. These changes improve code reusability and structural consistency across the project.