refactor: button component styles and variants
Made the button component styles' functions in both Angular and React components. Instead of directly assigning the result of the style functions, they now return functions that are later invoked in the markup files. Additionally, the variant condition in the Angular button style was changed from 'outlined' to 'elevated'. In the story, more button variants were added.