chore: refactor styles and structure for tabs component
This commit introduces multiple changes to the tabs component. The primary change is the modification of default and state-based class definitions in StylesHelper.ts. The code is refactored to only add a 'relative' class if the key is the default key. Additionally, in Tab.tsx, the code adjusts and cleans up the structure of classNames, ultimately improving element styling. Refactoring is also done in TabStyle.ts to further improve the flexibility of the tabs component.