feat: add 'name' prop and 'onChange' handler to Slider component
In the Slider component, a new 'name' property was added to the component props. An 'onChange' handler function was also introduced to update the value when it changes. Moreover, a hidden input field was added to store the name and value of the slider, which may be used when the form containing the slider needs to be submitted. The debug console.log that was printing the percentage value was also removed.