feat: add icon position property to Button component
The Button component has been updated to support an optional 'iconPosition' property. This new property allows to define the position of the icon within the button, if present; either on the 'left' or 'right'. The button styles and class names have also been adapted to this change for a proper layout.