feat: add ref to Tab event and implement scrolling
Multiple modifications are integrated to improve tab navigation. Firstly, a 'ref' element was added to the 'onTabSelected' event in Tab.tsx for better reference handling. Secondly, a scrolling behavior was implemented in Tabs.tsx. When a tab is selected, the tablist will automatically scroll to center the selected tab, enhancing user experience. This is particularly useful for a scrollable tab list.